backend/requirements.txt
fastapi[standard]
fastapi-cli
python-dotenv
aiohttp[speedups]
redis[hiredis]
